Fixed waitlist functionality in RegistrationManager
-enrollStudentInCourse() now only adds a course to a student's schedule
if the waitlist is empty after calling the enroll method i.e. the
student was added to the roll
-also simplified the canEnroll() method in CourseRoll. Now uses
contains() method to traverse the course roll to check for duplicate
students. Also now doesn't explicitly check for duplicates in the
waitlist queue, relies on the queues own inherent ability to now allow
duplicates